+++ source/LearningTests/LearningTests.csproj
@@ -43,4 +43,8 @@ limitations under the License.
       <None Remove="appsettings.json" />
     </ItemGroup>
 
+    <ItemGroup>
+      <Folder Include="Infrastructure" />
+    </ItemGroup>
+
 </Project>
