+++ source/Messaging.Infrastructure/Transactions/Aggregations/FakeGridAreaLookup.cs
@@ -17,7 +17,7 @@ using System.Threading.Tasks;
 using Messaging.Application.Transactions.Aggregations;
 using Messaging.Domain.Actors;
 
-namespace Messaging.Infrastructure.Transactions.AggregatedTimeSeries;
+namespace Messaging.Infrastructure.Transactions.Aggregations;
 
 public class FakeGridAreaLookup : IGridAreaLookup
 {
